Working Order History - Login functionality for Patrons and Admins
18/10/2024

Login works and displays orders IF the patron has completed a previous order.

Admin login is unchanged.

Added banner to warn user when logging in as an admin.

Added banner to profile page if user ends up on this page without being logged in.

Added error banner when user attempts to login and login is unsuccessful.